requirements.txt
accelerate==1.4.0
annotated-types==0.7.0
asttokens==2.4.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
dask==2024.8.0
debugpy==1.8.1
decorator==5.1.1
docstring-parser==0.16
emoji==2.14.1
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.17.0
fire==0.7.0
fsspec==2025.2.0
huggingface-hub==0.29.2
idna==3.10
importlib-metadata==8.6.1
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.0
jupyter-core==5.7.2
locket==1.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
mistral-common==1.5.3
mistral-inference==1.5.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
pandas==2.2.3
parso==0.8.4
partd==1.4.2
peft==0.14.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
platformdirs==4.2.2
prompt-toolkit==3.0.47
protobuf==6.30.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.1
pydantic==2.10.6
pydantic-core==2.27.2
Pygments==2.17.2
python-dateutil==2.9.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentencepiece==0.2.0
simple-parsing==0.1.7
six==1.16.0
stack-data==0.6.3
stanza==1.10.1
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
tomli==2.2.1
toolz==1.0.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.2
transformers==4.49.0
triton==3.2.0
typing-extensions>=4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
xformers==0.0.29.post3
zipp==3.21.0